more work on multiplayer

+++ src/elk/buildutil/ResourceSteps.hx
@@ -24,11 +24,9 @@ class ResourceSteps {
 		var build_files = sys.FileSystem.readDirectory(build_dir);
 		for (file in build_files) {
 			var file_path = new haxe.io.Path(sys.FileSystem.absolutePath('$build_dir/$file'));
-			if (sys.FileSystem.isDirectory(file_path.toString()))
-				continue;
+			if (sys.FileSystem.isDirectory(file_path.toString())) continue;
 
-			if (file_path.ext == 'pak')
-				sys.FileSystem.deleteFile(file_path.toString());
+			if (file_path.ext == 'pak') sys.FileSystem.deleteFile(file_path.toString());
 		}
 
 		return null;
